remove vbe add gop

+++ kernel/std/stdint.h
@@ -9,7 +9,7 @@ typedef signed long long int int64_t;
 typedef unsigned char uint8_t;
 typedef unsigned short int uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long long int uint64_t;
+typedef unsigned long int uint64_t;
 typedef long unsigned int size_t;
 typedef long signed int ssize_t;
 typedef __int128_t int128_t;
